+++ models/events.go
@@ -20,6 +20,7 @@ type Event interface {
 
 var types = map[string](func() Event){
 	"environment_created":  func() Event { return &EnvironmentCreated{} },
+	"environment_deleted":  func() Event { return &EnvironmentDeleted{} },
 	"environments_ordered": func() Event { return &EnvironmentsOrdered{} },
 	"feature_created":      func() Event { return &FeatureCreated{} },
 	"feature_toggled":      func() Event { return &FeatureToggled{} },